Fix string_pos according to old GM manuals

+++ ENIGMAsystem/SHELL/Universal_System/estring.h
@@ -38,23 +38,23 @@ size_t string_length(const char* str);
 #define string_byte_length(x) string_length(x)
 size_t string_length_utf8(string str);
 size_t string_length_utf8(const char* str);
-int string_pos(string substr,string str);
+size_t string_pos(string substr, string str);
 
 string string_format(double val, unsigned tot, unsigned dec);
-string string_copy(string str, int index,int count);
+string string_copy(string str, int index, int count);
 string string_set_byte_at(string str, int pos, char byte);
 char string_byte_at(string str, int index);
 string string_char_at(string str,int index);
-string string_delete(string str,int index,int count);
-string string_insert(string substr,string str,int index);
-string string_replace(string str,string substr,string newstr);
-string string_replace_all(string str,string substr,string newstr);
+string string_delete(string str, int index, int count);
+string string_insert(string substr, string str, int index);
+string string_replace(string str, string substr, string newstr);
+string string_replace_all(string str, string substr, string newstr);
 size_t string_count(string substr, string str);
 
 string string_lower(string str);
 string string_upper(string str);
 
-string string_repeat(string str,int count);
+string string_repeat(string str, int count);
 
 string string_letters(string str);
 string string_digits(string str);
